Include urban in "crops", to make sure the area involved in urban transitions is accounted for.
If urban is ignored, then the transitions to/from urban are missed,
and therefore the resulting fractions of all land use types diverge
from the data, in regions where urban transitions are present.

This is going to be revised when we merge urban code, and urban areas
become a separate land use type.