Implemented workaround for multiple OpenMP threads and debug mode crash.
There is apparent compiler issue, when with multiple openmp threads *and* debug
flags Intel compilers (15 and 16) report index errors in dump_diag_field_with_sel,
as if land_tile_map array started from 1,1 instead of is,js. Passing this array in
as an argument resolves this issue.

Credit to Zhi for finding this workaround and implementing it on his branch.