Add "protected" attribute (F2003) to most of the variables exported from the modules.
This should add modicum of safety to the export of data from the modules,
by protecting the exported data from modifications in the modules that use
them.

There are few variables that could not be made protected. For example, heat
capacities (cpw, clw, csw) are exported from soil_tile_mod, but they are actually
appear in the namelist in soil_mod. Such cases are is historical and could be
fixed if we were willing to re-arrange the namelists, a disruptive change.

Also, lnd%time is updated by the land model, therefore lnd could not be made
protected in land_data_mod.