Removed sizes of the dimensions from add_tile_data interfaces.
The interfaces of the add_tile_data subroutine are now the same as before
unstructured grid mods, but the way it determines the length of varoable's
dimensions is different. For the sake of efficiency, the netcdf file is no
longer queried: instead, code uses cached values of dimension sizes from
restart data structure.