Add min_lai threshold in vegn_data_nml to avoid numerical problems due to tiny LAI.
This is in response to the problem Huan encountered in her 1850 run,
where LAI was as low as 0.125e-305, causing crash due to numerics
in water terms derivative calculations.

If update_biomass_pools -- with light_saber turned on -- tries to bring
LAI below min_lai, bl (and therefore LAI) is set to zero.