Adjust initial psi_l in plant hydraulics calculations so that it is not stuck at very low values.
When the soil freezes, psi_r (and, consequently, psi_x and psi_l) are extremely low.
In previous code, such very low psi_l resulted in zero stomatal conductance that was
stuck at zero even when the soil was no longer frozen, because of the way initial
conditions for the implicit solution were set up.