Add by-species diagnostics for cohort diag fields.
Add "(species)" right after the name of the diag fields to get it by-species; no spaces,
and (I guess) it must be lowercase. Example:
"vegn",  "nindivs(species)",  "nindivs_sp", ......

The species along the species axis are in the same order as in the input field_table,
or in the species table in fms.out and logfile.