Save land cell_area diag field on structured grid.
This is the same as in LM4p0; it is necessary for re-gridding of land output,
since cell area must be known everywhere, not just over the grid cells where land
exists.

This implies chnages to diag table: cell_area must go into a separate file, because
structured and unstructured grid output cannot be mixed together in the same file.