Calculate scorch parameter as weighted average of canopy trees.
This will change answers if "unpacked" fire is used in PPA mode.

It introduces by-species parameter F_scorch_height, which was previously
hard-coded from Thonicke et al. (2010); now it needs to be specified
in field table for each tree species. Non-woody vegetation is ignored
in the calculation of average, so the value of this parameter does not
matter for grasses, currently.