Add grass height scaling factor for layer assignment.
Grasses bend under the wind, exposing tree seedlings to light even if the
seedlings are shorter; therefore we can scale the effective height of the
grasses down for the purposes of assigning layers to the vegetation cohorts.

This introduces new species parameter layer_height_factor; default is 1.0
to reproduce older code.