Add an option to remove all but natural vegetation tiles.
If there are no natural vegetation tiles (or their area is zero), then the tile with
highest vegetation biomass is designated as natural, and the rest is removed.

Areas of natural vegetation tiles is scaled up to occupy the entire soil in the grid cell.

This is primarily to convert 1850 land use state from piControls to potential vegetation
suitable for starting historic land use spin-ups. It is supposed to be used in a one-shot
runs, just to generate IC.