Moved updates of e_res_1 and e_res_2 outside of the canopy water and LW re-linearization loops in the fast time scale land solver.
Otherwise those residuals change values as we iterate to improve solution
of energy/water balance system, which is in contrast with how we treat other
components of energy/water balance equations, such as SW fluxes. This should
improve energy conservation properties of the solution.

Note that THIS UPDATE CHANGES ANSWERS.