Change irrigation fraction input
Irrigated crop fraction is very difficult to interpolate conservatively
and consistently to a model grid, because it needs to be weighted with
crop area. Instead, irrigated fraction of grid cells is calculated
offline in pre-processing and then read by the model, together with crop
fraction, to get the consistent fraction of croplands on model grid.

Also, refactor land transition code to simplify recalculating transitions
due to irrigation fraction change.