Reduce dinesionality of the linear optimization problem
Instead of considering transitions from/to every land use types,
we can aggregate the non-cropland (irrigated or rain-fed) types tugether,
calculate the transitions to/from rain-fed and irrigated croplands,
and then disaggregate the results to get the transition rates to/from
non-crop tiles using simple proportionality assumption. This simplifies
the irrigation transition splitting code rather significantly, and is likely
to make it less costly and more robust.

Also, to remove ambiguity in linear optimization, the cost of transition
between non-crop and irrigated areas is set a little higher than between
rain-feds and irrigated.