Use all LU types that model includes in crops to calculate irrigated crop fraction
The model currectly assigns urban areas to crop type, so that the area of
cities is included in the cropland area. However, when calculating irrigated
fraction of crops from input data, urban area was not taken into account,
resulting in irrigated area discrepancy that grew disproportionally to total
irrigated area with the increase of urbanization.