Calculate soil frac immediately before seed transport, instead of on initialization
This CHANGES ANSWERS at the time or vegetation reproduction [which is
currently the end of the year] if seed transport is on.

This should CROSS-RESTART FIX REPRODUCIBILITY ISSUE in seed transport.

The problem: soil fraction (m2 of soil per m2 of land) is used as an
area normalizing factor in seed transport calculations (in either
"spread" or "diffuse" options). Prior to this commit, soil fraction was
calculated on model initialization. Mathematically, soil fraction in the
model is a constant regardless of the circumstances, but because number
of sub-grid tiles changes as the model runs, and the soil frac
calculated as a sum of tile fractions, the calculated values are
slightly different from restart to restart, due to different
order-of-operations and finite machine precision. Therefore, in a
regression test where restart occurs between the start of the simulation
and the time of seed transport, this soil fraction normalizing factor is
tiny bit different from the test that ran continuously.

Those differences lead to non-reproducibility across restarts, when an
intermediate restart -- and therefore model initialization -- happens
between the start of the test and the time of seed transport
application.

SOLUTION: if soil frac is calculated immediately before the seed transport,
the order of tile fraction summation and the the resulting soil frac are
the same regardless of the intermediate restarts.

HOWEVER, this also means that this commit changes answers from the
previous version because of the differences in order of operation in
soil frac calculations.