Add missing root litter deposition in vegn_phenology_ppa
Root litter was missed when carbon was deposited to soil after leaf-off.
This would lead to violation of carbon conservation, except in this case
root litter is zero, because root_mort_rate happens to be hardcoded to
be 0 -- that is, root biomass does not decline after the leaf drop. This
fix does not change answers, but allows to use finite root lifetime --
if necessary -- in the future development of the model.

Tested by temporarily setting root_mort_rate = 0.1 [1/day] and running
3-month test of LM4p2scLitt-C96-coldstart. It does trigger conservation
check crash if root carbon is not taken into account, but with this fix
the test runs fine to completion.