Use add_soil_matter in CORPSE-specific code of vegn_graze_pasture_lm3
This may potentially change answers for specific set of options (namely,
daily grazing in lm4p0 mode with CORPSE soil carbon), because of the
intricacies of CORPSE soil BGC cohort creation/merging, and negative input
protection.

The benefit is that the debris from grazing would be included in litterfall
diagnostics from vegetation, and the CORPSE-specific interface is
minimised.