Handle negative inputs to litter pools, plus more sanity checks
Negative carbon inputs to litter can happen when, for example, carbon_gain
is negative and the fire deposits it on the ground. This update handles it
the same way CORPSE does: accumulate negatives in buffers (instead of adding
negatives to litter pools) and fill these buffers in with the positive
inputs as soon as possible.