Correct issues in DOC concentration updates after advection
There are still a question whether the current update method is correct:
Current treatment distributes resulting DOC between bulk and rhizosphere
proportionally to the values that existed before advection (that is,
scales previous values up or down).

It seems reasonable at the first glance, but does this treatment describe what
happens in advection correctly?

There should be some exchange between rhizosphere and bulk soil due to water flow,
in addition to diffusion that should also exist.

Also, numerically, what if the initial value of DOC is positive but very small?
Is it possible we get crazy numbers in this case?

The same questions apply to the updates of litter pools, except the DOC exchange
between pools is -- probably -- of less importance because they are more separated
than rhizosphere and bulk soil.