Fix a logical error in i/o of irrigation fraction for dust emission
If no dust tracers are defined in the field table, the dust module
initialization sets do_dust to FALSE and quits, without doing
initializing dust module internal variables, because there is
no point of doing so. But the irrigation fraction input code in the
dust module still tried to read the irrigation fraction -- calling
the subroutines that rely on some data structures initialized in
land_dust_init. Because they were not initialized in this case, the
model was failing at the end of the model year with signal 11
(SEGFAULT). Note that traceback did not give clear indication where
in the code the problem was happening.