Moved some vegetation radiative settings to new vegn_radiation_nml
This is to allow controlling these settings -- in particular intercepted
snow albedo -- independently from the settings used for the snowpack. It
is necessary in case of GLASS snow model with BRDF albedo, and will
allow more flexibility in general in the future.

NOTE that this code change will require change in namelists to reproduce
older results.