Change initialization of optimal layer distribution to avoid division by 0.
When depth-opt_layer_N is positive but below opt_layer(1)/2, arrays z
and l in dzopt structure were initialized with NaNs in production mode.
In debug mode, that condition was causing a crash.

As a consequence, in production mode, that should have led to thin layers
not being merged, because any comparison with NaN always results in FALSE.

This fix changes answers, since the thin layers can now be merged.