Avoid realloctaion of global arrays, for thread safety
iGlobal arrays opt_layer and opt_layer_z were used in the code of dzopt_t, a
derived type used to describe the optimal layer thickness distribution. At
some point, these arrays were changed from static to re-allocatable, and
their values were re-allocated and re-assigned depending on the
perceived need for array size. This was not thread-safe, e.g. when two
threads try to change these arrays simultaneously.

This update removes the need for these arrays.

It also introduces a new namelist parameter opt_layer_top in
snowpack_nml, allowing one to specify the optimal thickness of the top
snowpack layer. The default is set to 5cm as it was hard-coded in the
previous versions, for reproducibility of the results.