Call tracer update before snow%step2.
This is necessary to simplify passing the current value of interactive tracer
deposition to the snow.

Note that this changes answers because of the change in the order of operations.
The interactive deposition of tracers may use the snow fraction, which now comes
from the same time step, instead of the updated value in the original treatment.