Use generic soil_BGC_t interface to deposit litter in vegn_graze_pasture_ppa.
This is to allow the land use to work with GIMICS in PPA: implementation of that
mode was missing in the previous code.

NOTE that it may change the results in CORPSE mode, mostly because CORPSE
add_soil_matter handles negatives in an appropriate way, which the previous code
did not.