Switch to bit-reproducing way of calculating global sums in crop seed transport
This change addresses the reproducibility issue observed between C5 and C6
clusters in ESM4.5 tests conducted by Eric Stofferahn. Function mpp_sum used
previously may not be bit-reproducible if the number of processing elements
(PEs) or the domain decomposition changes. However, the mpp_global sum is
reproducible.

NOTE that this commit may change answers in the experiments with land use
(specifically, in the runs with non-zero croplands), because of the change
in order of operations.