Add ability to read CMIP7 land use data set
This adds plantations (pltns) to the list of land use types. Note that due to the lack
of any guidance on what to do with plantations, they are currently lumped together with
the secondary land use type. This also adds pltns_harv as a secondary-to-secondary
transition.

Also note that in the current version of the CMIP7 historical land use (v20250310), all
plantations and related transitions appear to be zero.