Use parameter from vegetation species table for dry matter caculations
This (together with other changes in the code) makes the fire emission
code independent of the number and order of the vegetation species in
the field table.

NOTE, that C_per_dry_matter parameter must be specified in the
vegetation species entries in the field table, requiring the changes
to the model configuration to reproduce old results.