Fix an issue with negative under-canopy precipitation
When vegetation dies, intercepted water/snow amounts and respective heat
contents are added to the "drop buffers" in the vegetation tile, to be added to
under-vegetation precipitation in the next time steps. The problem arises when
the linearized solution results in negative water on the leaves right before
vegetation dies. This negative water in some cases may result in negative "drop
buffer" content, and that in turn may produce negative below-canopy
precipitation. That, in turn, breaks the GLASS snow model which is sensitive to
such physically impossible input.

The solution is to handle "drop buffers" more carefully, only adding positive
mass to the precipitation, and borrowing the mass from precipitation to fill the
buffers.

NOTE that while we attempt to avoid negative canopy water/snow in linear solver,
the protection does not work perfectly in case of numerous cohorts -- it would
require multiple iterations to do that.

This will changeanswers in caseis when drop does go negative; it may change
the answers in other cases due to a change in the order of operations.