Avoid overheated water when distributing deficits in snow_sublimation
In some situations, the temperatures of the meltwater become unreasonably high;
this results in very high heat content of runoff (amplified by the ratio of soil
to lake areas) that the lake temperature jumps out of bounds, crashing the
model. In particular, this happens when the ws(1) + topsnowdeficit is a small
positive, but topsnowheatdeficit is so large that it is enough to melt all the
snow in the top layer, and then heat it up to super-high temperatures.

The solution this commit implements is to spend only a fraction of
topsnowheatdeficit, so that the meltwater temperature stays within a reasonable
range: the rest of topsnowheatdeficit remains, and can contribute to the next
time steps. It sounds reasonable to set the meltwater T to TFREEZE, although
admittedly it is a rather arbitrary choice.

This commit does change answers.