Reconcile the representation of transition matrix
It appears that a previous git merge has mixed up two different representations
of transition rate matrix: for each grid cell, one as the array of tran_type
(donor,acceptor,frac) structures, and the other as real frac(donor,acceptor), in
the part of the code that calculates and adds irrigation transitions based on
the areas and assumed costs.

The code in this commit uses the latter approach.