Add an option to match selected crops to the land use type
By default (FALSE), any crop (rain-fed or irrigated) can be selected for any
cropland land use type (irrigated or rain-fed). This is appropriate for the case
when irrigation transitions are not turned on in the land use module: all crop
tiles are tagged as "rain-fed" in this case, so excluding certain crops from
selection would likely result in underestimation of the crop coverage compared
to the real world.

The option match_crops_with_landuse_water_source = TRUE in vegn_crop_nml changes
this behavior, forcing the match: that is, rain-fed crops are not allowed on
irrigated land use tiles, and irrigated crops are not allowed on rain-fed tiles.