Harvest date must be less than a year from planting
Since the harvest date is encoded as a day-of-year, any interval larger than or
equal to 365 days will result in harvesting occurring very soon (or immediately)
after the planting.

Peter Phillipps had already fixed this issue earlier, but for some reason, it
didn’t make it into the latest version in the repository.