Use nonlinear qsat equation for fog threshold
In the original formulation, the saturated specific humidity required
to trigger fog formation was determined using a linearized equation
w.r.t canopy air temprature.

This linearized expression can easily yield negative values for the
fog threshold when the temperature tendency in a specific time step is
large in magnitude and negative. For instance, if we assume a 7%/K
temperature dependence of qsat, a roughly 14.3 K decrease in T will
result in negative values. In general, the linearized expression will
lead to smaller fog thresholds (compared to the precise equation for
qsat) and, consequently, more frequent fog formation in the conditions
where it should not.

Setting a new namelist parameter fog_qsat_fix in land_model_nml to
FALSE should revert to the original behavior.